+++ sentry-android-okhttp/src/main/java/io/sentry/android/okhttp/SentryOkHttpInterceptor.kt
@@ -66,8 +66,18 @@ class SentryOkHttpInterceptor(
         val url = urlDetails.urlOrFallback
         val method = request.method
 
-        // read transaction from the bound scope
-        val span = hub.span?.startChild("http.client", "$method $url")
+        val span: ISpan?
+        val isFromEventListener: Boolean
+
+        if (SentryOkHttpEventListener.eventMap.containsKey(chain.call())) {
+            // read the span from the event listener
+            span = SentryOkHttpEventListener.eventMap[chain.call()]?.callRootSpan
+            isFromEventListener = true
+        } else {
+            // read the span from the bound scope
+            span = hub.span?.startChild("http.client", "$method $url")
+            isFromEventListener = false
+        }
         urlDetails.applyToSpan(span)
 
         var response: Response? = null
@@ -105,38 +115,51 @@ class SentryOkHttpInterceptor(
             }
             throw e
         } finally {
-            finishSpan(span, request, response)
+            finishSpan(span, request, response, isFromEventListener)
 
-            val breadcrumb = Breadcrumb.http(request.url.toString(), request.method, code)
-            request.body?.contentLength().ifHasValidLength {
-                breadcrumb.setData("request_body_size", it)
+            // The SentryOkHttpEventListener will send the breadcrumb itself if used for this call
+            if (!isFromEventListener) {
+                sendBreadcrumb(request, code, response)
             }
-
-            val hint = Hint()
-                .also { it.set(OKHTTP_REQUEST, request) }
-            response?.let {
-                it.body?.contentLength().ifHasValidLength { responseBodySize ->
-                    breadcrumb.setData("response_body_size", responseBodySize)
-                }
-
-                hint[OKHTTP_RESPONSE] = it
-            }
-
-            hub.addBreadcrumb(breadcrumb, hint)
         }
     }
 
-    private fun finishSpan(span: ISpan?, request: Request, response: Response?) {
-        if (span != null) {
-            if (beforeSpan != null) {
-                val result = beforeSpan.execute(span, request, response)
-                if (result == null) {
-                    // span is dropped
-                    span.spanContext.sampled = false
-                } else {
+    private fun sendBreadcrumb(request: Request, code: Int?, response: Response?) {
+        val breadcrumb = Breadcrumb.http(request.url.toString(), request.method, code)
+        request.body?.contentLength().ifHasValidLength {
+            breadcrumb.setData("http.request_content_length", it)
+        }
+
+        val hint = Hint().also { it.set(OKHTTP_REQUEST, request) }
+        response?.let {
+            it.body?.contentLength().ifHasValidLength { responseBodySize ->
+                breadcrumb.setData("http.response_content_length", responseBodySize)
+            }
+
+            hint[OKHTTP_RESPONSE] = it
+        }
+
+        hub.addBreadcrumb(breadcrumb, hint)
+    }
+
+    private fun finishSpan(span: ISpan?, request: Request, response: Response?, isFromEventListener: Boolean) {
+        if (span == null) {
+            return
+        }
+        if (beforeSpan != null) {
+            val result = beforeSpan.execute(span, request, response)
+            if (result == null) {
+                // span is dropped
+                span.spanContext.sampled = false
+            } else {
+                // The SentryOkHttpEventListener will finish the span itself if used for this call
+                if (!isFromEventListener) {
                     span.finish()
                 }
-            } else {
+            }
+        } else {
+            // The SentryOkHttpEventListener will finish the span itself if used for this call
+            if (!isFromEventListener) {
                 span.finish()
             }
         }