+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -367,6 +367,7 @@ class SentryOptionsTest {
         externalOptions.addContextTag("requestId")
         externalOptions.proguardUuid = "1234"
         externalOptions.idleTimeout = 1500L
+        externalOptions.bundleIds.addAll(listOf("12ea7a02-46ac-44c0-a5bb-6d1fd9586411 ", " faa3ab42-b1bd-4659-af8e-1682324aa744"))
         val options = SentryOptions()
 
         options.merge(externalOptions)
@@ -390,6 +391,7 @@ class SentryOptionsTest {
         assertEquals(listOf("userId", "requestId"), options.contextTags)
         assertEquals("1234", options.proguardUuid)
         assertEquals(1500L, options.idleTimeout)
+        assertEquals(setOf("12ea7a02-46ac-44c0-a5bb-6d1fd9586411", "faa3ab42-b1bd-4659-af8e-1682324aa744"), options.bundleIds)
     }
 
     @Test